🌱 extract agent vocabulary leaves on v4
